requirements.txt
asgiref==3.8.1
Django==4.2.17
django-ckeditor-5==0.2.15
django-js-asset==2.2.0
djangorestframework==3.15.2
pillow==11.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
